refactor: move some more types to Github/Application
