+++ Source/Infrastructure/Integration/Kysect.Shreks.Integration.Github/Extensions/ServiceCollectionExtensions.cs
@@ -1,6 +1,7 @@
 using System.Security.Claims;
 using GitHubJwt;
 using Kysect.Shreks.Application.Abstractions.Github;
+using Kysect.Shreks.Integration.Github.Applicaiton;
 using Kysect.Shreks.Integration.Github.Client;
 using Kysect.Shreks.Integration.Github.CredentialStores;
 using Kysect.Shreks.Integration.Github.Entities;
